+++ lib/HeadersGeneration/HelpersToHeader.cpp
@@ -19,12 +19,12 @@
 #include "revng/HeadersGeneration/PTMLHeaderBuilder.h"
 #include "revng/Model/Binary.h"
 #include "revng/Model/FunctionTags.h"
+#include "revng/PTML/CAttributes.h"
 #include "revng/PTML/CBuilder.h"
 #include "revng/PTML/IndentedOstream.h"
 #include "revng/Pipeline/Location.h"
 #include "revng/Pipes/Ranks.h"
 #include "revng/Support/Assert.h"
-#include "revng/Support/CAttributes.h"
 #include "revng/Support/Debug.h"
 #include "revng/TypeNames/LLVMTypeNames.h"
