+++ lib/TypeNames/ModelTypeNames.cpp
@@ -22,13 +22,13 @@
 #include "revng/Model/Helpers.h"
 #include "revng/Model/PointerType.h"
 #include "revng/Model/RawFunctionDefinition.h"
+#include "revng/PTML/CAttributes.h"
 #include "revng/PTML/CBuilder.h"
 #include "revng/PTML/Constants.h"
 #include "revng/PTML/Tag.h"
 #include "revng/Pipeline/Location.h"
 #include "revng/Pipes/Ranks.h"
 #include "revng/Support/Assert.h"
-#include "revng/Support/CAttributes.h"
 #include "revng/TypeNames/LLVMTypeNames.h"
 #include "revng/TypeNames/ModelCBuilder.h"
